+++ tests/acceptance/user-click-on-test.js
@@ -16,15 +16,26 @@ test('button', assert => {
 
   andThen(function() {
     assert.equal(find('#submitted').length, 0, 'clicking button submits form');
-    assert.equal(find('#nested-submitted').length, 0, 'clicking button submits nested form');
   });
 
   clickOn('form.button');
-  clickOn('nested.button');
 
   andThen(function() {
     assert.ok(findWithAssert('#submitted'), 'clicking button submits form');
-    assert.ok(findWithAssert('#nested-submitted'), 'clicking button submits nested form');
+  });
+});
+
+test('anchor', assert => {
+  visit('/');
+
+  andThen(function() {
+    assert.equal(find('#submitted').length, 0, 'clicking anchor submits form');
+  });
+
+  clickOn('form.anchor');
+
+  andThen(function() {
+    assert.ok(findWithAssert('#submitted'), 'clicking anchor submits form');
   });
 });
 
@@ -42,6 +53,38 @@ test('submit', assert => {
   andThen(function() {
     assert.ok(findWithAssert('#submitted'), 'clicking button submits form');
     assert.ok(findWithAssert('#nested-submitted'), 'clicking button submits nested form');
+  });
+});
+
+test('radio', assert => {
+  visit('/');
+
+  clickOn('form.radio.yes');
+  clickOn('nested.radio.yes');
+
+  andThen(function() {
+    assert.ok(
+      findWithAssert('#yes').prop('checked'),
+      'clicking yes checks it'
+    );
+    assert.ok(
+      findWithAssert('#nested-yes').prop('checked'),
+      'clicking nested yes checks it'
+    );
+  });
+
+  clickOn('form.radio.no');
+  clickOn('nested.radio.no');
+
+  andThen(function() {
+    assert.ok(
+      findWithAssert('#no').prop('checked'),
+      'clicking no checks it'
+    );
+    assert.ok(
+      findWithAssert('#nested-no').prop('checked'),
+      'clicking nested no checks it'
+    );
   });
 });